add new method

+++ phoenix-knowledge-api/src/main/java/com/ginkgocap/ywxt/knowledge/service/KnowledgeService.java
@@ -267,5 +267,5 @@ public interface KnowledgeService {
 	 */
 	public InterfaceResult<List<DataCollection>> getBaseByColumnIdAndType(short columnId,short type,int start,int size) throws Exception;
 
-	public long getKnowledgeCount(long userId) throws Exception;
+	public int getKnowledgeCount(long userId) throws Exception;
 }
